+++ docs/CHANGELOG.md
@@ -10,6 +10,9 @@ This document holds the user-facing changelog that is also used in release notes
 ### Fixed
 - Fixed high CPU usage when the application is running in the background.
 - Fixed wrong tooltip styling when hovering over node icons in Light theme.
+- Fixed a bug where creating a new project would unintentionally clear the data in the Game Data Manager window.
+
+---
 
 ## v0.9-alpha - 2026-Jan-16
 - Initial release